requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
google-genai>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0
streamlit>=1.30.0
